This is my development "master" branch of librepcb.
I want the application be usable so I will manage my own build with my
modifications and it depends on U.Bruhin if he wants my modifications
in the main repo. He can tell "this thing I like" and I can prepare PR
for that, but I will not spent a lot of energy to something he denies
after while, or something he denies I can't touch. It can be usable for
him to see some version of work, we can discuss more options and
variants of solution.

I need user-experience usable apps and this app is not very usable for
me for now, as some user-experience things have not been fixed for years
and seems it will not be for next few years. There is no roadmap when
the UX refacoring will be done and any information against that.

Primarily I like to solve some UX problems in 1.0 version. This is my
"release" branch which contains all differences to the main branch in
the version 03d3930e. When root master move forward, I create my new
branch based on new root of parent as I can track modifications that
I was done. As there will be always merge commits, all of my
modifications are trackable.

I will rebase all of my feaure branches to the actual master, so it
contain always mergable code to be prepared for U.Bruhin if he like
to merge something upstream. Beware that any feature branches in my
fork will have moving SHA (rebasing workflow) as I like all things
I done be comparable with actual upstream master so don't stick to
SHA's. Moreover, they can be growing as some fixes there will be
needed.