Add SymbolMetadataDock
This commit is first in series with the intent to solve LibrePCB#815.

At first, it seemed that only adding of gripper will be needed, but after implementing, I revealed that
adding grippers as recommended by dbrgn changes UI file completely and there is no trace what has been
changed inside. From that point of view, i agree with Urban, that this change is very huge, needs to
spent a lot of time to implement and test so it will be probably never merged.

As far as LibrePCB#514 is must have for me (because of item selection), I can arrange series of commits
in the way it follows LibrePCB#726 (not the library but the movable tool dock widget). The package editor
is currently not usable for me as I can't switch the layers to show how the package looks like on different
outputs, AND I can't select right item on different layer when they overlap.

Another linked issue is LibrePCB#359 maximizes space for drawing as the metadata can be hidden and should not
distract user when drawing on small screens. This can be solved along LibrePCB#726

In LibrePCB#213 there is hint which way the UI will be changed in the future and from that point of view,
converting right tool window into dock widget will solve probably the future compatibility as in the
transition meantime the components can be partly QtQuick and partly old style QtWidgets, and some parts of
work can be pushed back on master by maintainer if he will like that.

So my modification is to introduce "editor docks" initially for metadata, that are shared by other editors.
This needs slight modification to context as the editor needs to access shared docks owned by window
and that will be probably exact reason to not accept my modifications into master by Urban.

Assumptions:

 * The Editor window is responsible for creating AND positioning of all docks.
 * There is new metadata editor for each widget so memory footprint of the dialog is minimized.
 * If the editor window is switched, the metadata is preloaded with new component.
 * TODO BUG: there should be check if there are modifications during leave !!!

So as result, I will try to refactor all tools to be separate objects to be compatible with all issues
mentioned above. I will rebase this branch periodically to be in sync with actual upstream master to be
compatible with Urban's. He can slap me for code style (i am not familiar with his coding style, I can
refactor this code to follow his style). Each commit will be compilable and staying separate so Urban
can grab some ideas from there.

BUG: When editor tabs are switched and there are modifications inside edit boxes, they are discarded now, no dirty
flag changes. Probably switch component needs to create commit before switch when dirty.