Refactor format_code.sh
There was plenty copy-pasted code inside format_code.sh. Adding new formatters
and options can introduce new bugs.

Proposed refactoring shortened all copied code to two bash function calls.

The search_files function searches for all intended files with specified extensions
they should be checked in the concrete linter.

The update_file checks, if the file content has been modified by linter tool and if
it's true, updates file and logs about the change on screen

The NEW_CONTENT is transferred to the update_file via global variable, but it can be
read from stdin until eof so the linter calls can be one-liners somewhere in the future,
but for now the reason is to keep the trace of changes from previous version of script.