Allow to specify branch in formatting script to compare against
In current code, only master branch is hardcoded as reference to compare and track changes against them.
This modifies lot of files when working on branch that are modified by other people in history when not
rooted on master.

I propose adding option "--compare-against <branch_name>" where you can select different branch (or tag or sha)
to compare local modifications against it.

There could be discussion about the argument name.

Probably there can be implementation detail where ALL and COMPARE_AGAINST have same meaning in the manner
of selecting which method to call, so in future the ALL variable can be left out and COMPARE_AGAINST will
be set to empty when "--all" argument is passed (KISS principle).